EM now runs but seems buggy
I fixed a lot of the bugs in the code and now it runs through fully. I have also implemented the restarts, but have not randomized the center points and variance. Something seems off with the results of the log likelihoods. They always come up to the same numbers within 2 runs.